snarkos360: moved locktick reporting to cli crate
