+++ src/lib.rs
@@ -16,8 +16,6 @@
 use amareleo_chain_cli::{commands::CLI, helpers::Updater};
 
 use clap::Parser;
-#[cfg(feature = "locktick")]
-use locktick::lock_snapshots;
 use std::{env, process::exit};
 
 #[cfg(all(target_os = "linux", target_arch = "x86_64"))]
@@ -39,36 +37,6 @@ pub struct BuildInfo<'a> {
 pub fn main_core(build: &BuildInfo) -> anyhow::Result<()> {
     // A hack to avoid having to go through clap to display advanced version information.
     check_for_version(build);
-
-    #[cfg(feature = "locktick")]
-    std::thread::spawn(|| {
-        loop {
-            tracing::info!("[locktick] checking for active lock guards");
-            let mut infos = lock_snapshots();
-            infos.sort_unstable_by(|l1, l2| l1.location.cmp(&l2.location));
-
-            for lock in infos {
-                let mut guards = lock.known_guards.values().collect::<Vec<_>>();
-                guards.sort_unstable_by(|g1, g2| g1.location.cmp(&g2.location));
-
-                for guard in guards.iter().filter(|g| g.num_active_uses() != 0) {
-                    let location = &guard.location;
-                    let kind = guard.kind;
-                    let num_uses = guard.num_uses;
-                    let active_users = guard.num_active_uses();
-                    let avg_duration = guard.avg_duration();
-                    let avg_wait_time = guard.avg_wait_time();
-                    tracing::info!(
-                        "{location} ({:?}): {num_uses}; {active_users} active; avg d: {:?}; avg w: {:?}",
-                        kind,
-                        avg_duration,
-                        avg_wait_time
-                    );
-                }
-            }
-            std::thread::sleep(std::time::Duration::from_secs(3));
-        }
-    });
 
     // Parse the given arguments.
     let cli = CLI::parse();
